refactor: auth 폴더 정리
모든 가입은 기본적으로 일반 USER 로 ROLE 설정됩니다.
