+++ haeun/itemstore/src/main/java/com/meyame/itemstore/controller/AuthController.java
@@ -6,8 +6,8 @@ import com.meyame.itemstore.dto.response.auth.JwtAccessTokenResDto;
 import com.meyame.itemstore.dto.response.auth.JwtRefreshTokenResDto;
 import com.meyame.itemstore.dto.response.auth.SignUpResDto;
 import com.meyame.itemstore.jwt.JwtTokenProvider;
-import com.meyame.itemstore.service.GoogleLoginService;
-import com.meyame.itemstore.service.LocalLoginService;
+import com.meyame.itemstore.service.auth.GoogleLoginService;
+import com.meyame.itemstore.service.auth.LocalLoginService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
