+++ haeun/itemstore/src/main/java/com/meyame/itemstore/controller/UserController.java
@@ -4,8 +4,8 @@ import com.meyame.itemstore.dto.request.user.UserUpdateReqDto;
 import com.meyame.itemstore.dto.response.auth.UserInfoDto;
 import com.meyame.itemstore.dto.response.user.UserInfoResDto;
 import com.meyame.itemstore.dto.response.user.UserUpdateResDto;
-import com.meyame.itemstore.service.GoogleLoginService;
-import com.meyame.itemstore.service.LocalLoginService;
+import com.meyame.itemstore.service.auth.GoogleLoginService;
+import com.meyame.itemstore.service.auth.LocalLoginService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
